pyproject.toml
[project]
name = "nonebot-plugin-twitter"
version = "0.0.1"
description = "NoneBot2 plugin for twitter"
authors = [
    {name = "nek0us", email = "nekouss@gmail.com"},
]
license = {text = "MIT"}
dependencies = ["nonebot2>=2.0.0", "nonebot-adapter-onebot>=2.1.3", "httpx>=0.23.0", "tweepy>=4.14.0" ,"nonebot_plugin_apscheduler>=0.2.0"]
requires-python = ">=3.8"
readme = "README.md"

[project.urls]
Homepage = "https://github.com/nek0us/nonebot-plugin-twitter"
Repository = "https://github.com/nek0us/nonebot-plugin-twitter"

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
